requirements.txt
# Core Dependencies
ollama>=0.2.0
httpx>=0.25.0
aiohttp>=3.9.0

# LangChain (optional, for extended features)
langchain>=0.1.0
langchain-community>=0.0.10

# Data Processing
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI & Display
rich>=13.0.0
click>=8.1.0
typer>=0.9.0

# Web UI
streamlit>=1.30.0
streamlit-extras>=0.3.0

# Data Analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Database & Persistence
aiosqlite>=0.19.0
sqlalchemy>=2.0.0

# Async & Performance
asyncio-throttle>=1.0.0
tenacity>=8.2.0
cachetools>=5.3.0

# Logging
loguru>=0.7.0
structlog>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
hypothesis>=6.90.0

# Optional: Alternative Backends
vllm>=0.2.0
llama-cpp-python>=0.2.0
transformers>=4.35.0
torch>=2.1.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0
ruff>=0.1.0
pre-commit>=3.5.0
